+++ lib/src/models/media_info_functions.dart
@@ -2,137 +2,137 @@ import 'dart:ffi';
 
 import 'package:ffi/ffi.dart';
 
-//void*           __stdcall MediaInfoA_New ()
-typedef MediaInfoANew = Pointer<Void> Function();
-typedef MediaInfoInit = Pointer<Void> Function();
+//void*           __stdcall MediaInfo_New ()
+typedef MediaInfoNew = Pointer<Void> Function();
+typedef MediaInfoInitFT = Pointer<Void> Function();
 
-//void*           __stdcall MediaInfoA_New_Quick (const char* File, const char* Options)
-typedef MediaInfoANewQuick = Pointer<Void> Function(
-  Pointer<Utf8> file,
-  Pointer<Utf8> options,
-);
+//void*           __stdcall MediaInfo_New_Quick (const char_t* File, const char_t* Options)
 typedef MediaInfoNewQuick = Pointer<Void> Function(
-  Pointer<Utf8> file,
-  Pointer<Utf8> options,
+  Pointer<WChar> file,
+  Pointer<WChar> options,
+);
+typedef MediaInfoNewQuickFT = Pointer<Void> Function(
+  Pointer<WChar> file,
+  Pointer<WChar> options,
 );
 
-//void            __stdcall MediaInfoA_Delete (void* Handle)
-typedef MediaInfoADelete = Void Function(
+//void            __stdcall MediaInfo_Delete (void* Handle)
+typedef MediaInfoDelete = Void Function(
   Pointer<Void> handle,
 );
-typedef MediaInfoDelete = void Function(
+typedef MediaInfoDeleteFT = void Function(
   Pointer<Void> handle,
-);
-
-//size_t          __stdcall MediaInfoA_Open (void* Handle, const char* File)
-typedef MediaInfoAOpen = Uint32 Function(
-  Pointer<Void> handle,
-  Pointer<Utf8> file,
-);
-typedef MediaInfoOpen = int Function(
-  Pointer<Void> handle,
-  Pointer<Utf8> file,
 );
 
-//size_t          __stdcall MediaInfoA_Open_Buffer (void* Handle, const unsigned char* Begin, size_t  Begin_Size, const unsigned char* End, size_t  End_Size)
-typedef MediaInfoAOpenBuffer = Uint32 Function(
+//size_t          __stdcall MediaInfo_Open (void* Handle, const wchar_t* File)
+typedef MediaInfoOpen = Size Function(
   Pointer<Void> handle,
-  Pointer<Utf8> begin,
+  Pointer<WChar> file,
+);
+typedef MediaInfoOpenFT = int Function(
+  Pointer<Void> handle,
+  Pointer<WChar> file,
+);
+
+//size_t          __stdcall MediaInfo_Open_Buffer (void* Handle, const unsigned wchar_t* Begin, size_t  Begin_Size, const unsigned wchar_t* End, size_t  End_Size)
+typedef MediaInfoOpenBuffer = Uint32 Function(
+  Pointer<Void> handle,
+  Pointer<WChar> begin,
   Uint32 beginSize,
-  Pointer<Utf8> end,
+  Pointer<WChar> end,
   Uint32 endSize,
 );
-typedef MediaInfoOpenBuffer = int Function(
+typedef MediaInfoOpenBufferFT = int Function(
   Pointer<Void> handle,
-  Pointer<Utf8> begin,
+  Pointer<WChar> begin,
   int beginSize,
-  Pointer<Utf8> end,
+  Pointer<WChar> end,
   int endSize,
 );
 
-//size_t          __stdcall MediaInfoA_Open_Buffer_Init (void* Handle, MediaInfo_int64u File_Size, MediaInfo_int64u File_Offset)
-typedef MediaInfoAOpenBufferInit = Uint32 Function(
+//size_t          __stdcall MediaInfo_Open_Buffer_Init (void* Handle, MediaInfo_int64u File_Size, MediaInfo_int64u File_Offset)
+typedef MediaInfoOpenBufferInit = Uint32 Function(
   Pointer<Void> handle,
   Uint64 fileSize,
   Uint64 fileOffset,
 );
-typedef MediaInfoOpenBufferInit = int Function(
+typedef MediaInfoOpenBufferInitFT = int Function(
   Pointer<Void> handle,
   int fileSize,
   int fileOffset,
 );
 
-//size_t          __stdcall MediaInfoA_Open_Buffer_Continue (void* Handle, MediaInfo_int8u* Buffer, size_t Buffer_Size)
-typedef MediaInfoAOpenBufferContinue = Uint32 Function(
+//size_t          __stdcall MediaInfo_Open_Buffer_Continue (void* Handle, MediaInfo_int8u* Buffer, size_t Buffer_Size)
+typedef MediaInfoOpenBufferContinue = Uint32 Function(
   Pointer<Void> handle,
   Pointer<Uint8> buffer,
   Uint32 bufferSize,
 );
-typedef MediaInfoOpenBufferContinue = int Function(
+typedef MediaInfoOpenBufferContinueFT = int Function(
   Pointer<Void> handle,
   Pointer<Uint8> buffer,
   int bufferSize,
 );
 
-//MediaInfo_int64u __stdcall MediaInfoA_Open_Buffer_Continue_GoTo_Get (void* Handle)
-typedef MediaInfoAOpenBufferContinueGoToGet = Uint64 Function(
+//MediaInfo_int64u __stdcall MediaInfo_Open_Buffer_Continue_GoTo_Get (void* Handle)
+typedef MediaInfoOpenBufferContinueGoToGet = Uint64 Function(
   Pointer<Void> handle,
 );
-typedef MediaInfoOpenBufferContinueGoToGet = int Function(
-  Pointer<Void> handle,
-);
-
-//size_t          __stdcall MediaInfoA_Open_Buffer_Finalize (void* Handle)
-typedef MediaInfoAOpenBufferFinalize = Uint32 Function(
-  Pointer<Void> handle,
-);
-typedef MediaInfoOpenBufferFinalize = int Function(
+typedef MediaInfoOpenBufferContinueGoToGetFT = int Function(
   Pointer<Void> handle,
 );
 
-//size_t          __stdcall MediaInfoA_Open_NextPacket (void* Handle)
-typedef MediaInfoAOpenNextPacket = Uint32 Function(
+//size_t          __stdcall MediaInfo_Open_Buffer_Finalize (void* Handle)
+typedef MediaInfoOpenBufferFinalize = Uint32 Function(
   Pointer<Void> handle,
 );
-typedef MediaInfoOpenNextPacket = int Function(
-  Pointer<Void> handle,
-);
-
-//size_t          __stdcall MediaInfoA_Save (void* Handle)
-typedef MediaInfoASave = Uint32 Function(
-  Pointer<Void> handle,
-);
-typedef MediaInfoSave = int Function(
+typedef MediaInfoOpenBufferFinalizeFT = int Function(
   Pointer<Void> handle,
 );
 
-//void            __stdcall MediaInfoA_Close (void* Handle)
-typedef MediaInfoAClose = Void Function(
+//size_t          __stdcall MediaInfo_Open_NextPacket (void* Handle)
+typedef MediaInfoOpenNextPacket = Uint32 Function(
   Pointer<Void> handle,
 );
-typedef MediaInfoClose = void Function(
+typedef MediaInfoOpenNextPacketFT = int Function(
   Pointer<Void> handle,
 );
 
-//const char*     __stdcall MediaInfoA_Inform (void* Handle, size_t Reserved)
-typedef MediaInfoAInform = Pointer<Utf8> Function(
+//size_t          __stdcall MediaInfo_Save (void* Handle)
+typedef MediaInfoSave = Uint32 Function(
+  Pointer<Void> handle,
+);
+typedef MediaInfoSaveFT = int Function(
+  Pointer<Void> handle,
+);
+
+//void            __stdcall MediaInfo_Close (void* Handle)
+typedef MediaInfoClose = Void Function(
+  Pointer<Void> handle,
+);
+typedef MediaInfoCloseFT = void Function(
+  Pointer<Void> handle,
+);
+
+//const wchar_t*     __stdcall MediaInfo_Inform (void* Handle, size_t Reserved)
+typedef MediaInfoInform = Pointer<WChar> Function(
   Pointer<Void> handle,
   Uint32 reserved,
 );
-typedef MediaInfoInform = Pointer<Utf8> Function(
+typedef MediaInfoInformFT = Pointer<WChar> Function(
   Pointer<Void> handle,
   int reserved,
 );
 
-//const char*     __stdcall MediaInfoA_GetI (void* Handle, MediaInfo_stream_t StreamKind, size_t StreamNumber, size_t  Parameter, MediaInfo_info_C KindOfInfo)
-typedef MediaInfoAGetI = Pointer<Utf8> Function(
+//const wchar_t*     __stdcall MediaInfo_GetI (void* Handle, MediaInfo_stream_t StreamKind, size_t StreamNumber, size_t  Parameter, MediaInfo_info_C KindOfInfo)
+typedef MediaInfoGetI = Pointer<WChar> Function(
   Pointer<Void> handle,
   Uint32 streamKind,
   Uint32 streamNumber,
   Uint32 parameter,
   Uint32 kindOfInfo,
 );
-typedef MediaInfoGetI = Pointer<Utf8> Function(
+typedef MediaInfoGetIFT = Pointer<WChar> Function(
   Pointer<Void> handle,
   int streamKind,
   int streamNumber,
@@ -140,108 +140,115 @@ typedef MediaInfoGetI = Pointer<Utf8> Function(
   int kindOfInfo,
 );
 
-//const char*     __stdcall MediaInfoA_Get (void* Handle, MediaInfo_stream_t StreamKind, size_t StreamNumber, const char* Parameter, MediaInfo_info_C KindOfInfo, MediaInfo_info_C KindOfSearch)
-typedef MediaInfoAGet = Pointer<Utf8> Function(
+//const wchar_t*     __stdcall MediaInfo_Get (void* Handle, MediaInfo_stream_t StreamKind, size_t StreamNumber, const wchar_t* Parameter, MediaInfo_info_C KindOfInfo, MediaInfo_info_C KindOfSearch)
+typedef MediaInfoGet = Pointer<WChar> Function(
   Pointer<Void> handle,
   Uint32 streamKind,
   Uint32 streamNumber,
-  Pointer<Utf8> parameter,
+  Pointer<WChar> parameter,
   Uint32 kindOfInfo,
   Uint32 kindOfSearch,
 );
-typedef MediaInfoGet = Pointer<Utf8> Function(
+typedef MediaInfoGetFT = Pointer<WChar> Function(
   Pointer<Void> handle,
   int streamKind,
   int streamNumber,
-  Pointer<Utf8> parameter,
+  Pointer<WChar> parameter,
   int kindOfInfo,
   int kindOfSearch,
 );
 
-//size_t          __stdcall MediaInfoA_SetI (void* Handle, const char* ToSet, MediaInfo_stream_t StreamKind, size_t StreamNumber, size_t  Parameter, const char* OldParameter)
-typedef MediaInfoASetI = Uint32 Function(
+//size_t          __stdcall MediaInfo_SetI (void* Handle, const wchar_t* ToSet, MediaInfo_stream_t StreamKind, size_t StreamNumber, size_t  Parameter, const wchar_t* OldParameter)
+typedef MediaInfoSetI = Uint32 Function(
   Pointer<Void> handle,
-  Pointer<Utf8> toSet,
+  Pointer<WChar> toSet,
   Uint32 streamKind,
   Uint32 streamNumber,
   Uint32 parameter,
-  Pointer<Utf8> oldParameter,
+  Pointer<WChar> oldParameter,
 );
-typedef MediaInfoSetI = int Function(
+typedef MediaInfoSetIFT = int Function(
   Pointer<Void> handle,
-  Pointer<Utf8> toSet,
+  Pointer<WChar> toSet,
   int streamKind,
   int streamNumber,
   int parameter,
-  Pointer<Utf8> oldParameter,
+  Pointer<WChar> oldParameter,
 );
 
-//size_t          __stdcall MediaInfoA_Set (void* Handle, const char* ToSet, MediaInfo_stream_t StreamKind, size_t StreamNumber, const char* Parameter, const char* OldParameter)
-typedef MediaInfoASet = Uint32 Function(
+//size_t          __stdcall MediaInfo_Set (void* Handle, const wchar_t* ToSet, MediaInfo_stream_t StreamKind, size_t StreamNumber, const wchar_t* Parameter, const wchar_t* OldParameter)
+typedef MediaInfoSet = Uint32 Function(
   Pointer<Void> handle,
-  Pointer<Utf8> toSet,
+  Pointer<WChar> toSet,
   Uint32 streamKind,
   Uint32 streamNumber,
-  Pointer<Utf8> parameter,
-  Pointer<Utf8> oldParameter,
+  Pointer<WChar> parameter,
+  Pointer<WChar> oldParameter,
 );
-typedef MediaInfoSet = int Function(
+typedef MediaInfoSetFT = int Function(
   Pointer<Void> handle,
-  Pointer<Utf8> toSet,
+  Pointer<WChar> toSet,
   int streamKind,
   int streamNumber,
-  Pointer<Utf8> parameter,
-  Pointer<Utf8> oldParameter,
+  Pointer<WChar> parameter,
+  Pointer<WChar> oldParameter,
 );
 
-//size_t          __stdcall MediaInfoA_Output_Buffer_Get (void* Handle, const char* Value)
-typedef MediaInfoAOutputBufferGet = Uint32 Function(
+//size_t          __stdcall MediaInfo_Output_Buffer_Get (void* Handle, const wchar_t* Value)
+typedef MediaInfoOutputBufferGet = Uint32 Function(
   Pointer<Void> handle,
-  Pointer<Utf8> value,
+  Pointer<WChar> value,
 );
-typedef MediaInfoOutputBufferGet = int Function(
+typedef MediaInfoOutputBufferGetFT = int Function(
   Pointer<Void> handle,
-  Pointer<Utf8> value,
+  Pointer<WChar> value,
 );
 
-//size_t          __stdcall MediaInfoA_Output_Buffer_GetI (void* Handle, size_t Pos)
-typedef MediaInfoAOutputBufferGetI = Uint32 Function(
+//size_t          __stdcall MediaInfo_Output_Buffer_GetI (void* Handle, size_t Pos)
+typedef MediaInfoOutputBufferGetI = Uint32 Function(
   Pointer<Void> handle,
   Uint32 int,
 );
-typedef MediaInfoOutputBufferGetI = int Function(
+typedef MediaInfoOutputBufferGetIFT = int Function(
   Pointer<Void> handle,
   int int,
 );
 
-//const char*     __stdcall MediaInfoA_Option (void* Handle, const char* Option, const char* Value)
-typedef MediaInfoAOption = Pointer<Utf8> Function(
+//const wchar_t*     __stdcall MediaInfo_Option (void* Handle, const wchar_t* Option, const wchar_t* Value)
+typedef MediaInfoOption = Pointer<WChar> Function(
   Pointer<Void> handle,
-  Pointer<Utf8> option,
-  Pointer<Utf8> value,
+  Pointer<WChar> option,
+  Pointer<WChar> value,
 );
-typedef MediaInfoOption = Pointer<Utf8> Function(
+typedef MediaInfoOptionFT = Pointer<WChar> Function(
   Pointer<Void> handle,
-  Pointer<Utf8> option,
-  Pointer<Utf8> value,
-);
-
-//size_t          __stdcall MediaInfoA_State_Get(void* Handle)
-typedef MediaInfoAStateGet = Uint32 Function(
-  Pointer<Void> handle,
-);
-typedef MediaInfoStateGet = int Function(
-  Pointer<Void> handle,
+  Pointer<WChar> option,
+  Pointer<WChar> value,
 );
 
-//size_t          __stdcall MediaInfoA_Count_Get(void* Handle, MediaInfo_stream_t StreamKind, size_t StreamNumber)
-typedef MediaInfoACountGet = Uint32 Function(
+//size_t          __stdcall MediaInfo_State_Get(void* Handle)
+typedef MediaInfoStateGet = Uint32 Function(
+  Pointer<Void> handle,
+);
+typedef MediaInfoStateGetFT = int Function(
+  Pointer<Void> handle,
+);
+
+//size_t          __stdcall MediaInfo_Count_Get(void* Handle, MediaInfo_stream_t StreamKind, size_t StreamNumber)
+typedef MediaInfoCountGet = Uint32 Function(
   Pointer<Void> handle,
   Uint32 streamKind,
   Uint32 streamNumber,
 );
-typedef MediaInfoCountGet = int Function(
+typedef MediaInfoCountGetFT = int Function(
   Pointer<Void> handle,
   int streamKind,
   int streamNumber,
+);
+
+
+typedef OpenFileForMediaInfo = int Function(
+  Pointer<Utf8> myDlibPath,
+  Pointer<Void> handle,
+  Pointer<Utf8> path,
 );