+++ test/mediainfo_test.dart
@@ -7,77 +7,117 @@ import "package:path/path.dart" as path;
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
-
-  late String _debugPath;
+  
+  late String debugPath;
 
   if (Platform.isLinux) {
-    _debugPath = "linux/libs/";
+    debugPath = "linux/libs";
   }
   if (Platform.isMacOS) {
-    _debugPath = "macos/libs/";
+    debugPath = "macos/libs";
   }
   if (Platform.isWindows) {
-    _debugPath = "windows/libs/";
+    debugPath = "windows/libs";
   }
 
   test('load mediainfo', () async {
-    Mediainfo.init(customDebugPath: _debugPath);
+    Mediainfo(customDebugPath: debugPath);
   });
 
   test('mediainfo get version', () async {
-    final _mi = Mediainfo.init(customDebugPath: _debugPath);
-    final optionResult = _mi.option("Info_Version");
+    final mi = Mediainfo(customDebugPath: debugPath);
+    final optionResult = mi.option("Info_Version");
 
-    expect(optionResult, "MediaInfoLib - v21.09");
+    expect(optionResult, "MediaInfoLib - v22.09");
+  });
+
+  test('mediainfo file inform', () async {
+    final musicPath = path.join(Directory.current.path, "assets/Beat_Thee.mp3");
+    final mi = Mediainfo(customDebugPath: debugPath);
+    mi.quickLoad(musicPath);
+    final optionResult = mi.inform();
+
+    expect(optionResult.isNotEmpty, true);
   });
 
   test('Get music data with quick load', () async {
     final musicPath = path.join(Directory.current.path, "assets/Beat_Thee.mp3");
-    final _mi = Mediainfo.init(customDebugPath: _debugPath);
-    _mi.quickLoad(musicPath);
+    final mi = Mediainfo(customDebugPath: debugPath);
+    mi.quickLoad(musicPath);
     expect(
-        _mi.getInfo(
+        mi.getInfo(
             MediaInfoStreamType.mediaInfoStreamAudio, 0, "BitRate/String"),
         "320 kb/s");
     expect(
-        _mi.getInfo(
+        mi.getInfo(
             MediaInfoStreamType.mediaInfoStreamAudio, 0, "Duration/String2"),
         "3 min 11 s");
     expect(
-        _mi.getInfo(
+        mi.getInfo(
             MediaInfoStreamType.mediaInfoStreamAudio, 0, "SamplingRate/String"),
         "44.1 kHz");
     expect(
-        _mi.getInfo(
+        mi.getInfo(
             MediaInfoStreamType.mediaInfoStreamAudio, 0, "FrameRate/String"),
         "38.281 FPS (1152 SPF)");
 
-    _mi.delete();
+    mi.delete();
+  });
+
+  test('Get music file with unicode symbols on quick load', () async {
+    final musicPath = path.join(Directory.current.path, "assets/テスト音楽.mp3");
+    final mi = Mediainfo(customDebugPath: debugPath);
+    mi.quickLoad(musicPath);
+    expect(
+        mi.getInfo(
+            MediaInfoStreamType.mediaInfoStreamAudio, 0, "BitRate/String"),
+        "320 kb/s");
+    expect(
+        mi.getInfo(
+            MediaInfoStreamType.mediaInfoStreamAudio, 0, "Duration/String2"),
+        "3 min 11 s");
+    expect(
+        mi.getInfo(
+            MediaInfoStreamType.mediaInfoStreamAudio, 0, "SamplingRate/String"),
+        "44.1 kHz");
+    expect(
+        mi.getInfo(
+            MediaInfoStreamType.mediaInfoStreamAudio, 0, "FrameRate/String"),
+        "38.281 FPS (1152 SPF)");
+
+    mi.delete();
   });
 
   test('Get music data with open', () async {
     final musicPath = path.join(Directory.current.path, "assets/Beat_Thee.mp3");
-    final _mi = Mediainfo.init(customDebugPath: _debugPath);
-    _mi.init();
-    _mi.open(musicPath);
+    final mi = Mediainfo(customDebugPath: debugPath);
+    mi.init();
+    mi.open(musicPath);
     expect(
-        _mi.getInfo(
+        mi.getInfo(
             MediaInfoStreamType.mediaInfoStreamAudio, 0, "BitRate/String"),
         "320 kb/s");
     expect(
-        _mi.getInfo(
+        mi.getInfo(
             MediaInfoStreamType.mediaInfoStreamAudio, 0, "Duration/String2"),
         "3 min 11 s");
     expect(
-        _mi.getInfo(
+        mi.getInfo(
             MediaInfoStreamType.mediaInfoStreamAudio, 0, "SamplingRate/String"),
         "44.1 kHz");
     expect(
-        _mi.getInfo(
+        mi.getInfo(
             MediaInfoStreamType.mediaInfoStreamAudio, 0, "FrameRate/String"),
         "38.281 FPS (1152 SPF)");
 
-    _mi.close();
-    _mi.delete();
+    mi.close();
+    mi.delete();
+  });
+
+  test('Invalid usage close', () async {
+    final mi = Mediainfo(customDebugPath: debugPath);
+
+    mi.close();
+    mi.delete();
   });
 }