+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.2
+
+* Fixed copy and load dylib for macos.
+* Fixed example for macos.
+* Updated Flutter to 3.3.7
+* Updated libMediaInfo from v21.09 to v22.09
+* Changed exception name from MediaInfoInstanceHasExist to MediaInfoInstanceHasExists
+* Fixed open files with unicode symbols in path
+
 ## 0.0.1+1
 
 * Updated Readme.md